reject reason and detail view

+++ Data/Product.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string ProductName { get; set; } = "";
-        public decimal Price { get; set; } 
+        public int Price { get; set; } 
         public int Quantity { get; set; }
 
         public string Action { get; set; }
@@ -12,7 +12,7 @@
     public class ProductViewModel
     {
         public string ProductName { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public int Quantity { get; set; }
